Fix bug in README documentation.
The example used for opening port 80 for http had an inconsistency. The name of the file (port_http.erb) did not match the call to it (previously 'iptables_rule "http"'). This would result in an error where Chef said it couldn't find http.erb in any of the likely places.

Fixed the example to use 'iptables_rule "port_http"' instead.